clear ram package to str
